Fix gisgraphy
