+++ src/main/java/com/graphhopper/converter/api/GisgraphyAddressEntry.java
@@ -17,7 +17,7 @@ public class GisgraphyAddressEntry {
     private long sourceId;
 
     private String countryCode;
-    
+
     private String country;
 
     private String city;
@@ -35,6 +35,8 @@ public class GisgraphyAddressEntry {
     private String formatedPostal;
 
     private String geocodingLevel;
+
+    private String adm3Name;
 
     public long getId() {
         return id;
@@ -146,8 +148,8 @@ public class GisgraphyAddressEntry {
     public String getCountry() {
         return country;
     }
-    
-    public void setCountry(String country){
+
+    public void setCountry(String country) {
         this.country = country;
     }
 
@@ -157,6 +159,14 @@ public class GisgraphyAddressEntry {
 
     public void setGeocodingLevel(String geocodingLevel) {
         this.geocodingLevel = geocodingLevel;
+    }
+
+    public String getAdm3Name() {
+        return adm3Name;
+    }
+
+    public void setAdm3Name(String adm3Name) {
+        this.adm3Name = adm3Name;
     }
 
 }
